+++ lib/next_rails.rb
@@ -5,6 +5,7 @@ require "next_rails/version"
 require "next_rails/init"
 require "next_rails/bundle_report"
 require "next_rails/bundle_report/ruby_version_compatibility"
+require "next_rails/bundle_report/rails_version_compatibility"
 require "deprecation_tracker"
 
 module NextRails